+++ app/src/test/java/org/simple/clinic/editpatient/EditPatientInitTest.kt
@@ -23,7 +23,15 @@ class EditPatientInitTest {
   @Test
   fun `when screen is created, then load initial data`() {
     val patientUuid = UUID.fromString("e40f42f4-0867-4891-ac77-95df5fe1fdef")
-    val defaultModel = EditPatientModel.from(patient, patientAddress, patientPhoneNumber, dateOfBirthFormat, bangladeshNationalId, EditPatientState.NOT_SAVING_PATIENT)
+    val defaultModel = EditPatientModel.from(
+      patient,
+      patientAddress,
+      patientPhoneNumber,
+      dateOfBirthFormat,
+      bangladeshNationalId,
+      EditPatientState.NOT_SAVING_PATIENT,
+      false
+    )
 
     initSpec.whenInit(defaultModel).then(assertThatFirst(
         hasModel(defaultModel),
@@ -38,7 +46,15 @@ class EditPatientInitTest {
   @Test
   fun `when screen is created and village type ahead is not enabled, then do not fetch colony or villages`() {
     val patientUuid = UUID.fromString("e40f42f4-0867-4891-ac77-95df5fe1fdef")
-    val defaultModel = EditPatientModel.from(patient, patientAddress, patientPhoneNumber, dateOfBirthFormat, bangladeshNationalId, EditPatientState.NOT_SAVING_PATIENT)
+    val defaultModel = EditPatientModel.from(
+      patient,
+      patientAddress,
+      patientPhoneNumber,
+      dateOfBirthFormat,
+      bangladeshNationalId,
+      EditPatientState.NOT_SAVING_PATIENT,
+      false
+    )
     val initSpec = InitSpec(EditPatientInit(patient = patient,
         isVillageTypeAheadEnabled = false))
 
@@ -55,7 +71,15 @@ class EditPatientInitTest {
   fun `when screen is restored, then don't fetch colony or villages`() {
     val colonyOrVillages = listOf("Colony1", "Colony2", "Colony3", "Colony4")
 
-    val updatedVillageOrColonyNamesModel = EditPatientModel.from(patient, patientAddress, patientPhoneNumber, dateOfBirthFormat, bangladeshNationalId, EditPatientState.NOT_SAVING_PATIENT).updateColonyOrVillagesList(colonyOrVillages)
+    val updatedVillageOrColonyNamesModel = EditPatientModel.from(
+      patient,
+      patientAddress,
+      patientPhoneNumber,
+      dateOfBirthFormat,
+      bangladeshNationalId,
+      EditPatientState.NOT_SAVING_PATIENT,
+      false
+    ).updateColonyOrVillagesList(colonyOrVillages)
 
     initSpec.whenInit(updatedVillageOrColonyNamesModel).then(assertThatFirst(
         hasModel(updatedVillageOrColonyNamesModel),