+++ CHANGELOG.md
@@ -25,6 +25,9 @@
 - Run renovate bot daily after 12 am
 - Update next appointment card UI specifications
 - Set toolbar title in `ScanSimpleIdScreen` based on where it's opened from
+- Implement adding `NHID` in `EditPatientScreen`
+- Request camera permission when add nhid button is clicked in `EditPatientScreen`
+- Fix when scanned QR code error is shown the progress state continues to show
 - [In Progress: 11 Jan 2022] Implement adding NHID in `PatientEditScreen`
 - [In Progress: 17 Jan 2022] Implement adding scan support for BpPassport in `EditPatientScreen`